cpufreq: add Nightmare and Darkness CPU governors
Nightmare based on PegasusQ governor, but less aggressive and more stable.
This governor is a good compromise between performance and battery. In
addition to the SoD is a prevention because it usually does not hotplug.

Darkness based on Nightmare governor but more simple and fast, basic configs
but very complex structure. It is an updated version of the nightmare gov,
so far it is quite stable in tests

Based on commits:
 https://github.com/Slim80/Imperium_Kernel_TW_5.0.1/commit/510868b802e65540afbeb8dcd4e103c2b6b53b22
 https://github.com/Slim80/Imperium_Kernel_TW_5.0.1/commit/81bc6c6aa46b7b5784eab8dc249006d9199922e7